+++ src/pages/Signin/Signin.tsx
@@ -7,7 +7,7 @@ import { InputForm, lightText, Link, navigationTitle } from "layouts/Authenticat
 import GoogleButton from "react-google-button"
 import { getGoogleAuthUrl, signInWithEmailAndPassword } from "api/auth"
 import { Link as MuiLink, TextField } from "@mui/material"
-import { RoundedButton } from "elements/button"
+import { RoundedButton } from "elements/Button"
 import { ERoutes } from "routes"
 import { isValidEmail } from "utils"
