+++ src/routes/index.tsx
@@ -1,9 +1,9 @@
 import React, { ReactNode } from "react"
 import AuthenticateLayout from "layouts/AuthenticateLayout"
-import Signin from "pages/Signin"
-import Signup from "pages/Signup"
-import Dashboard from "pages/Dashboard"
-import Redirecting from "pages/Redirecting"
+import Signin from "pages/Signin/Signin"
+import Signup from "pages/Signup/Signup"
+import Dashboard from "pages/Dashboard/Dashboard"
+import Redirecting from "pages/Redirecting/Redirecting"
 import { Navigate, Route } from "react-router-dom"
 import MainScreen from "layouts/MainScreen"
 import { PrivateRoutes } from "utils/PrivateRoute"
